Added shorthand for all the commands
There are several commands that can be used to mark tasks as done, delete tasks, and add new tasks. These commands can be abbreviated to save time and make the commands more concise. For example, instead of typing "mark" and "delete", you can simply type "m" and "d" respectively.

For example, if you want to mark a task as done, you can type "m 1" instead of "mark 1". Similarly, you can use "d" to delete a task.

The "list" command can be used to display the list of tasks in the database. This command can be abbreviated to "l".

The "todo" command can be used to add a new todo task. This command can be abbreviated to "t".

The "deadline" command can be used to add a new deadline task. This command can be abbreviated to "d".

The "event" command can be used to add a new event task. This command can be abbreviated to "e".

This will make it easier for the user to manage their tasks and stay organized.